requirements.txt
Flask==2.3.3
google-genai==0.3.0
python-dotenv==1.0.0
requests==2.31.0
youtube-transcript-api==0.6.2